+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## v0.7.0 (2025-03-17)
 
 - Inline script step to reduce external network calls. See !186.
+- Upgrade Go to use version `1.24`. See !193.
 
 ## v0.6.0 (2025-02-24)